Move creation of log file to Inner. Get rid of BufWriterFactory
